requirements.txt
asgiref==3.8.1
Django==5.2.1
pillow==11.2.1
sqlparse==0.5.3
djangorestframework~=3.16.0


pytest~=8.3.5
drf-spectacular~=0.28.0
drf-spectacular~=0.28.0
rest_framework_simplejwt
djangorestframework~=3.16.0